Reorganize `path` module
