+++ src/pandss/paths/__init__.py
@@ -1,0 +1,3 @@
+from .dataset_path import DatasetPath
+from .dataset_path_collection import DatasetPathCollection
+from .wildcard import WILDCARD_PATTERN
