+++ src/pandss/paths/wildcard.py
@@ -1,0 +1,3 @@
+from re import compile
+
+WILDCARD_PATTERN = compile(r"\.\*")
